move BlifParser test logic into FOEDAG unittests section
